+++ apps/server/src/features/accounts/handlers/get-account-handler.ts
@@ -2,7 +2,7 @@ import { db } from '../../../db-config';
 import { AuthenticatedExpressRequest } from '../../../types/express';
 import { createHandler } from '../../../utils/create-handler';
 
-export const getAccount = createHandler(async (req: AuthenticatedExpressRequest, res) => {
+export const getAccountHandler = createHandler(async (req: AuthenticatedExpressRequest, res) => {
   const session = req.session;
   const acc = await db('account')
     .where({ user_id: session.user.id })