+++ apps/server/src/features/currencies/currencies-router.ts
@@ -1,10 +1,10 @@
 import { getRouter } from '../../utils/get-router';
 import { checkAuth } from '../auth/middleware/check-auth';
-import { getCirculation } from './handlers/get-circulation';
-import { getUserTokens } from './handlers/get-user-tokens';
+import { getCirculationHandler } from './handlers/get-circulation-handler';
+import { getUserTokensHandler } from './handlers/get-user-tokens-handler';
 
 const router = getRouter();
 
-router.get('/circulation', getCirculation);
-router.get('/tokens', checkAuth(), getUserTokens);
+router.get('/circulation', getCirculationHandler);
+router.get('/tokens', checkAuth(), getUserTokensHandler);
 export { router as currenciesRouter };