+++ apps/server/src/features/currencies/handlers/get-user-tokens-handler.ts
@@ -1,18 +1,13 @@
 import { db } from '../../../db-config';
 import { tablenames } from '../../../tablenames';
 import { AuthenticatedExpressRequest } from '../../../types/express';
+import { tokenRepo } from '../../../utils/classes/token-repo';
 import { createHandler } from '../../../utils/create-handler';
 import { getTokens } from '../helpers/get-tokens';
 
 /**Returns the tokens owned by the currently authenticated user. */
-export const getUserTokens = createHandler(async (req: AuthenticatedExpressRequest, res) => {
+export const getUserTokensHandler = createHandler(async (req: AuthenticatedExpressRequest, res) => {
   const session = req.session;
-  const tokens = await getTokens(db).where({
-    account_id: db
-      .select('id')
-      .from(tablenames.accounts)
-      .where({ user_id: session.user.id })
-      .limit(1),
-  });
+  const tokens = await tokenRepo.getTokensOfUserById(session.user.id, db);
   return res.status(200).json(tokens);
 });